DOXYGEN: Let doxygen once more look at the whole doxygen dir
At some point we stopped doing this, because various backends
pulled in tons of garbage symbols. However, manually adding
all subdirs under backends/ that we want is error prone and
requires a lot of manual intervention (e.g. base-backend.h/.cpp
and modular-backend.h/.cpp were forgotten until now).

So I figured it is time to try this once more. If a backend
pulls in too much unwanted stuff, we could add it to the
EXCLUDE list instead.

svn path=/doxygen/; revision=56111