BUILDBOT: Initial version of configuration update to build tools.
This looks good, and no changes to scumm.py code should be needed, but
it should be reviewed before it is implemented on the buildbot box.

Some changes may be needed to the tools builder code to remove
configuration options which are ScummVM only. Also, some work may be
needed to allow correct cross-compilation of tools and out of tree
builds.