BUILDBOT: Switch Triggerable schedulers to using change filters.
The previous usage of "branch" is deprecated for Triggerable schedulers
and this also prevents commits to the main repository triggering the
tools builds and vice versa i.e. previously both were commits to
"master" branches, but with different repository names.

FYI: The Nightly schedulers still use the "branch" parameters and in
fact require it according to the manual, so don't remove these.