BUILDBOT: Fix Dreamcast build with serial debug disabled - Try #2.
Silly mistake. The p structure is used by reference, not value so
appending new parameters to the same p structure caused both Dreamcast
builds to have the same parameters. This is not the entire issue as
the libronin-noserial was not picked up in this case, which is
incorrect. Will need a bit more work.