BUILDBOT: Add support code for changing archive format type per build.
This is to allow the packaged build's archive format i.e. currently
fixed at tar.bz to be varied per platform. Some platforms i.e.
Android would be much easier for these builds to be packaged in zip
format and this code adds most of the required code to support this
change.