BUILDBOT: Add code to allow build-only builders, skipping packaging.
This is intended for builders present to allow build compile
testing for specific compilers / configurations, rather than to generate
nightly packages for user/development testing.

This is important as the buildbot box has a limited amount of disk space
and each new builder which packages uses a significant amount for
storage of these. Have set this on the existing builders of this type.