BUILDBOT: Prevent "Clean" buildstep being killed by timeout.
The Clean buildstep has started failing as the "rm" command runs
with no output, and due to the size of the builders directories. after
1200 secs of no output, the buildbot framework assumes that the command
has got stuck and kills the process and fails.

This is preventing nightlies from running correctly.

To prevent this, the verbose flag is added to the rm command to ensure
output and prevent the timeout.