BUILDBOT: Avoid using failing debug symbol utility in new IOS toolchain.
The dsymutil-macosx is a OSX binary loaded via maloader and ld-mac,
which uses mmap() to do this. This currently fails, possibly as it
requires write permissions to dsymutil-macosx to achieve this.

Will need to investigate and avoid adding write permissions to other
to fix as this is highly insecure.

It may be possible to patch the ld-mac / maloader to fix or worst case,
change the dsymutil-macosx to be owned by buildbot user and set user
write permission.