BUILDBOT: Disable OS X PPC builder.
Static builds for OS X PPC failed since a long time due to binary size issues.
We added workarounds for this in 2014 already. But the situation got worse due
to even more engines being added in the meantime.
We tried adding gcc's -mlongcall option which is precisely for this case.
However, this causes internal compiler errors when compiling ScummVM.

Additionally, it looks like SDL was built with X11 support, but we do not have
any X11 development headers. This also causes compilation to fail.

Until we fix our toolchain issues we simply disable our builder.