BUILDBOT: Fix Compilation of Dingux Build.
Disable Theora decoder as the low memory branch of tremor used by the
Dingux toolchain does not support all the features required by the
Theora decoder.

If this is fixed in future, this can be re-enabled, but it is also
likely that any games using Theora video i.e. Broken Sword 2.5 would
not run on Dingux due to CPU and screen size limitations, so this may be
a moot point.