BUILDBOT: disable newer engines on PSP2 to fix generated executables.
There seems to be an ongoing issue for PS Vita that when newer engines
are added and thus the static binary gets larger, it fails to load and
run correctly. The "correct" solution to this would probably be to
implement plugin loading, but for now will disable some engines to allow
users to use the exectuables and test.