Implemented a JSONFormatter and WebHookReporter
The WebHook reporter builds on the default Notifier
that's also used for e-mails. This brings more configurability
for the messages that get send.

The WebHookReporter needs to be able to post JSON.
So a JSON formatter was born. The formatter is specific
to how Discord parses messages, but can be used
as an example for other backends.

Unfortunately the JSONFormatter needs to have multiple functions,
or else the message won't get created and send.