BUILDBOT: Disable the DockTile plugin for 64 bits macOS builds
This revert commit 31d3a27. The issue is that this build uses the
macOS 10.14 SDK (which doesn't support 32 bit builds) but targets
OX X 10.9. As our configure scripts uses the targeted system as
a proxy for the SDK version, it thinks 32 buts builds are possible.

If we want to enable back the dock tile plugin for this target we
will need one of two things: either a configure flag to force
targetting 64 bit only, or check the SDK version instead of using
the target OS version to set the 64 bits only flag.