Split target configuration from code.
The game targets are now specified in JSON and not in code.

This change makes it easier to add new game targets and
to run one game with different flags.

The Target dataclass is populated from a `targets.json` file
which follows this structure:
[
  {
    "name": "Spaceship Warlock",
    "directory": "warlock-win",
    "game_id": "warlock-win",
    "platform": "win",
    "version": "D3",
    "debugflags": "fewframesonly,fast",
    "movienames": ["moviename", ....]
 },
 {etc,...}
]

More information can be found in the readme.