Switch to reporting based on latest master
The goal is to make it easier to see the changes
based on master and not based on a previous run on another
branch or PullRequest.

For example when the last run was for a wobbly PR, it would
make a lot of tests as failed. Running a build on master
would make all those tests as fixed. Even though that commit
on master didn't touch those code paths.