Add logging on if change is director related
PR 4546 [1] had director related changes. Somehow a build wasn't
kicked off. Logging showed that a buildrequest wasn't injected
after calling `file_is_director_related`.

The working hypothesis is the `Change` didn't include the files.
To test this logging is added to show all files and whether or
not a build was triggered for it.

[1] https://github.com/scummvm/scummvm/pull/4546