CLOUD: Add new flow for 2.7.1 and higher
New flow displays the page that does cross-origin request, instead of displaying a short code that could be used to get the data from cloud.scummvm.org. The logic of storing and retrieving data from Redis is not needed for this flow, and could be removed in the future.

To switch between the flows, a checkbox to the index page was added.